plot_util/requirements.txt
humanfriendly==10.0
matplotlib==3.7.0
numpy==1.24.2
python-dotenv==0.21.1
